// angular/src/app/modules/sos/shared/models/Cartera.ts
import { DocCartera } from "./Doccartera";

export class Cartera {
    
    tipo:string;
    nombretd: string;
    numnitcar: string;
    fecfaccar: Date;
    codigotd:string;
    prefijocar: string;
    docucar: string;
    vrfaccar: number;
    vrpagcar: number;
    vence: Date;
    tm1car: string;
    docauxcar: string;
    nombrecli: string;
    digitocli: string;
    direccioncli: string;
    telefonocli: string;
    mora:number;
    dias:number;
    salcre: number;
    cuota:string;
    cuerpo: Array<DocCartera>;
}