// angular/src/app/modules/sos/shared/models/DocAuxCartera.ts
export class DocAuxCartera {
    
    tipo: string;
    nombretd: string;
    fefaccar: Date;
    codigotd:string;
    prefijocar: string;
    docucar: string;
    vrfaccar: string;
    vrpagcar: string;
    vence: Date;
    numnitcar: string;
    nombrecli: string;
    digitocli: string;
    direccioncli: string;
    telefonocli: string;
    nombreciu: string;
    debito: number;
    credito: number;
    saldo: number;
    mora: number;
    diast: number;
    salcre: string;
    cuota: string;
}